+++ molviewspec/molviewspec/builder.py
@@ -1037,6 +1037,25 @@ class Component(_Base, _FocusMixin, _TransformMixin):
     def representation(
         self,
         *,
+        type: Literal["backbone"],
+        size_factor: float | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Representation:
+        """
+        Add a backbone representation for this component.
+        :param type: the type of this representation ('backbone')
+        :param size_factor: adjust the scale of the visuals (relative to 1.0)
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: a builder that handles operations at representation level
+        """
+        ...
+
+    @overload
+    def representation(
+        self,
+        *,
         type: Literal["ball_and_stick"],
         ignore_hydrogens: bool | None = None,
         size_factor: float | None = None,
@@ -1046,6 +1065,27 @@ class Component(_Base, _FocusMixin, _TransformMixin):
         """
         Add a ball-and-stick representation for this component.
         :param type: the type of this representation ('ball_and_stick')
+        :param ignore_hydrogens: draw hydrogen atoms?
+        :param size_factor: adjust the scale of the visuals (relative to 1.0)
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: a builder that handles operations at representation level
+        """
+        ...
+
+    @overload
+    def representation(
+        self,
+        *,
+        type: Literal["line"],
+        ignore_hydrogens: bool | None = None,
+        size_factor: float | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Representation:
+        """
+        Add a line representation for this component.
+        :param type: the type of this representation ('line')
         :param ignore_hydrogens: draw hydrogen atoms?
         :param size_factor: adjust the scale of the visuals (relative to 1.0)
         :param custom: optional, custom data to attach to this node